+++ server/TWS Foundation/CSM Foundation/Database/Quality/Disposing/BQ_Disposer.cs
@@ -84,8 +84,8 @@ public abstract class BQ_Disposer
 
                 entry.DetectChanges();
                 entry.State = EntityState.Deleted;
+                database.SaveChanges();
             }
-            database.SaveChanges();
         }
     }
 }